Refactored out AbstractPagingProvider
No need for the parent class now that we only have "read". Also lazily instantiating a Transformer now so it's only created if there's an XML properties fragment to process.
